Adding src to search.

+++ src/Models/Network.cpp
@@ -1,6 +1,6 @@
 #include "Network.hpp"
 
-Network::Network(String ssid, float rssi, int encryptionType) {
+Network::Network(String ssid, int rssi, int encryptionType) {
   this->ssid = ssid;
   this->rssi = rssi;
   this->encryptionType = encryptionType;
